requirements.txt
# Good News Scraper & Slider Generator - Dependencies
# ================================================================

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# AI & API Integration
openai>=1.3.0
requests>=2.31.0

# Configuration & Environment
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0

# Web Scraping (if needed for future enhancements)
beautifulsoup4>=4.12.0

# Optional: For development and testing
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# isort>=5.12.0
